Use extend function to bind callback data to plot objects
This just makes the code cleaner. The ajax call returns some data to the
callback function, which passes it to the appropriate constructor
for a plot object. This changes the plot constructors so that they loop
over the ajax data and bind each field to a corresponding field on the
constructed object, rather than writing out each field by hand.